feat(ink): the WebGL2 renderer, the page buffer, and the ink host
The worker owns an OffscreenCanvas and draws every stroke as SDF capsules
in one instanced call, with the live stroke on its own layer. The page
buffer keeps strokes in a growable typed array with tombstone erase, the
stroke index answers near-point and near-segment queries for the eraser,
and shape-snap fits lines, rectangles, ellipses and arrows. The InkHost
mounts the canvas, the pen gate, and the bar.

+++ apps/web/src/features/ink/test/use-pen-capture.test.ts
@@ -255,7 +255,7 @@ test("a 240 Hz stroke through the session touches the screen once, at the end", 
   harness.session.committed(
     { strokeId: 1, pageIndex: 0, width: 6, tool: "pen", colour: "text" },
     new Float32Array([0, 0, 10, 10]),
-    new Float32Array([128, 128]),
+    new Uint8Array([128, 128]),
   );
   assert.equal(harness.committed.length, 1, "one acknowledgement, one render");
 });
